pyproject.toml
[project]
name = "protonvpn-portforwarding-qbittorrent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "py-natpmp>=0.2.5",
    "qbittorrent-api>=2025.5.0",
]

[dependency-groups]
dev = [
    "ruff>=0.11.12",
]
